pyproject.toml
[tool.poetry]
name = "quickstart-api-openai"
version = "0.1.0"
description = ""
authors = ["mielesantos <mielesnts@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.51.2"
python-dotenv = "^1.0.1"
groq = "^0.11.0"
langchain-openai = "^0.2.2"
langchain-community = "^0.3.2"
pymupdf = "^1.24.11"
duckduckgo-search = "^6.3.1"
langchain-experimental = "^0.3.2"
wikipedia = "^1.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
taskipy = "^1.12.2"
ruff = "^0.4.8"
httpx = "^0.27.0"

[tool.ruff]
line-length = 88
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check .&& ruff check . --diff'
format = 'ruff check . --fix; ruff format .'
run = 'python manage.py runserver'
pre_test = 'task lint'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"